Extract svg runtime component into a separate file
