+++ lib/loaders/svg-sprite-loader/svg-runtime-generator.js
@@ -1,3 +1,5 @@
+const fs = require('fs');
+const path = require('path');
 const { stringifyRequest } = require('loader-utils');
 const { stringifySymbol } = require('svg-sprite-loader/lib/utils');
 
@@ -7,37 +9,12 @@ const runtimeGenerator = ({ symbol, config, context }) => {
   const spriteRequest = stringifyRequest({ context }, spriteModule);
   const symbolRequest = stringifyRequest({ context }, symbolModule);
 
-  return `
-    import React from 'react';
-    import SpriteSymbol from ${symbolRequest};
-    import sprite from ${spriteRequest};
+  const component = fs.readFileSync(path.resolve(__dirname, 'component.js')).toString();
 
-    const symbol = new SpriteSymbol(${stringifySymbol(symbol)});
-    sprite.add(symbol);
-
-    const SvgSpriteIcon = function SvgSpriteIcon(props) {
-      return React.createElement(
-        'svg',
-        Object.assign({
-          viewBox: symbol.viewBox,
-        }, props),
-        React.createElement(
-          'use',
-          {
-            xlinkHref: '#' + symbol.id,
-          }
-        )
-      );
-    };
-
-    SvgSpriteIcon.viewBox = symbol.viewBox;
-    SvgSpriteIcon.id = symbol.id;
-    SvgSpriteIcon.content = symbol.content;
-    SvgSpriteIcon.url = symbol.url;
-    SvgSpriteIcon.toString = symbol.toString;
-
-    export default SvgSpriteIcon;
-  `;
+  return component
+    .replace('\'$$symbolRequest$$\'', symbolRequest)
+    .replace('\'$$spriteRequest$$\'', spriteRequest)
+    .replace('$$stringifiedSymbol$$', stringifySymbol(symbol));
 };
 
 module.exports = runtimeGenerator;
